+++ lib/rspec/mocks/verifying_proxy.rb
@@ -132,7 +132,7 @@ module RSpec
 
       def validate_arguments!(actual_args)
         @method_reference.with_signature do |signature|
-          verifier = Support::MethodSignatureVerifier.new(signature, actual_args)
+          verifier = Support::StrictSignatureVerifier.new(signature, actual_args)
           unless verifier.valid?
             raise ArgumentError, verifier.error_message
           end